ISIS-2726: improves docs for UserService
removes UserServiceDefault
